BumpyBeginnings/bumpybeginnings/notifications/serializers.py
from rest_framework import serializers
from .models import Notification

class NotificationSerializer(serializers.ModelSerializer):
    class Meta:
        model = Notification
        fields = ['id', 'recipient', 'notification_message', 'timestamp', 'is_read']
        read_only_fields = ['timestamp']