BumpyBeginnings/bumpybeginnings/requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
bleach==5.0.1
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.3
daphne==4.1.2
Django==4.2.16
django-allauth==65.2.0
django-bleach==3.1.0
django-ckeditor==6.7.1
django-environ==0.11.2
django-js-asset==2.2.0
django-rest-swagger==2.2.0
django-sslserver==0.22
django-tinymce==4.1.0
djangorestframework==3.15.2
drf-yasg==1.21.9
exceptiongroup==1.2.2
hyperlink==21.0.0
hypothesis==6.123.2
idna==3.10
incremental==24.7.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
msgpack==1.1.0
openapi-codec==1.3.2
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.0
pyOpenSSL==25.0.0
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
service-identity==24.2.0
simplejson==3.20.1
six==1.17.0
sortedcontainers==2.4.0
sqlparse==0.5.2
tinycss2==1.1.1
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
webencodings==0.5.1
zope.interface==7.2
